requirements.txt
flask
requests
pydantic
python-dotenv
